+++ Easy/NumberOf1Bits/tests/test.cpp
@@ -8,8 +8,8 @@ protected:
 TEST_F(NumberOf1Bits, FirstTest) {
     uint32_t n = 0b00000000000000000000000000001011;
 
-    int32_t result = Solution::hamming_weight(n);
-    int32_t expected = 3;
+    size_t result = Solution::hamming_weight(n);
+    size_t expected = 3;
 
     ASSERT_EQ(result, expected);
 }
@@ -17,8 +17,8 @@ TEST_F(NumberOf1Bits, FirstTest) {
 TEST_F(NumberOf1Bits, SecondTest) {
     uint32_t n = 0b00000000000000000000000010000000;
 
-    int32_t result = Solution::hamming_weight(n);
-    int32_t expected = 1;
+    size_t result = Solution::hamming_weight(n);
+    size_t expected = 1;
 
     ASSERT_EQ(result, expected);
 }
@@ -26,8 +26,8 @@ TEST_F(NumberOf1Bits, SecondTest) {
 TEST_F(NumberOf1Bits, ThirdTest) {
     uint32_t n = 0b11111111111111111111111111111101;
 
-    int32_t result = Solution::hamming_weight(n);
-    int32_t expected = 31;
+    size_t result = Solution::hamming_weight(n);
+    size_t expected = 31;
 
     ASSERT_EQ(result, expected);
 }
